refactor: simplify inline editor
